+++ fanboi2/resources/app/initialize.js
@@ -1,4 +1,5 @@
 domready(function(){
     require('components/navbar_toggler');
     require('components/quote_popover');
+    require('components/reply_fill');
 });